Move Add Item to a modal
